requirements.txt
torch>=1.9.0
transformers>=4.11.3
django>=3.2.8
djangorestframework>=3.12.4
numpy>=1.21.2
matplotlib>=3.4.3
